# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CSE341 Project 2 API",
        "version": "1.0.0",
        "description": "API for managing contacts and user profiles with full CRUD operations.",
        "contact": {
            "name": "CSE341 Course",
            "url": "https://www.byupattheway.org"
        },
        "license": {
            "name": "ISC"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development Server"
        },
        {
            "url": "https://your-render-app.onrender.com",
            "description": "Production Server (Render)"
        }
    ],
    "paths": {
        "/api/contacts": {
            "get": {
                "summary": "Retrieve all contacts",
                "description": "Fetches a list of all contacts in the database",
                "operationId": "getAllContacts",
                "tags": ["Contacts"],
                "responses": {
                    "200": {
                        "description": "A list of contacts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/Contact" }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error retrieving contacts",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new contact",
                "description": "Creates a new contact with all required fields.",
                "operationId": "createContact",
                "tags": ["Contacts"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/ContactInput" }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contact created successfully",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/Contact" } }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ValidationError" } }
                        }
                    },
                    "500": {
                        "description": "Server error creating contact",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
                        }
                    }
                }
            }
        },
        "/api/contacts/{id}": {
            "get": {
                "summary": "Retrieve a contact by ID",
                "description": "Fetches a single contact by its MongoDB _id",
                "operationId": "getContactById",
                "tags": ["Contacts"],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "description": "MongoDB _id", "schema": { "type": "string" } }
                ],
                "responses": {
                    "200": {
                        "description": "Contact found",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Contact" } } }
                    },
                    "400": { "description": "Invalid ID format" },
                    "404": { "description": "Contact not found" },
                    "500": { "description": "Server error" }
                }
            },
            "put": {
                "summary": "Update a contact by ID",
                "description": "Updates one or more fields of an existing contact",
                "operationId": "updateContact",
                "tags": ["Contacts"],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "description": "MongoDB _id", "schema": { "type": "string" } }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "firstName": { "type": "string" },
                                    "lastName": { "type": "string" },
                                    "email": { "type": "string", "format": "email" },
                                    "favoriteColor": { "type": "string" },
                                    "birthday": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contact updated",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Contact" } } }
                    },
                    "400": { "description": "Validation error or invalid ID" },
                    "404": { "description": "Contact not found" },
                    "500": { "description": "Server error" }
                }
            },
            "delete": {
                "summary": "Delete a contact by ID",
                "description": "Permanently removes a contact from the database",
                "operationId": "deleteContact",
                "tags": ["Contacts"],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "description": "MongoDB _id", "schema": { "type": "string" } }
                ],
                "responses": {
                    "200": {
                        "description": "Contact deleted",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object", "properties": { "message": { "type": "string" } } }
                            }
                        }
                    },
                    "400": { "description": "Invalid ID format" },
                    "404": { "description": "Contact not found" },
                    "500": { "description": "Server error" }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Contact": {
                "type": "object",
                "required": ["firstName", "lastName", "email", "favoriteColor", "birthday"],
                "properties": {
                    "_id": { "type": "string", "description": "MongoDB ObjectId" },
                    "firstName": { "type": "string", "description": "Contact first name" },
                    "lastName": { "type": "string", "description": "Contact last name" },
                    "email": { "type": "string", "format": "email", "description": "Contact email address" },
                    "favoriteColor": { "type": "string", "description": "Contact favorite color" },
                    "birthday": { "type": "string", "description": "Contact birthday" }
                }
            },
            "ContactInput": {
                "type": "object",
                "required": ["firstName", "lastName", "email", "favoriteColor", "birthday"],
                "properties": {
                    "firstName": { "type": "string" },
                    "lastName": { "type": "string" },
                    "email": { "type": "string", "format": "email" },
                    "favoriteColor": { "type": "string" },
                    "birthday": { "type": "string" }
                }
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "message": { "type": "string" },
                    "error": { "type": "string" }
                }
            },
            "ValidationError": {
                "type": "object",
                "properties": {
                    "message": { "type": "string" },
                    "errors": { "type": "array", "items": { "type": "string" } }
                }
            }
        }
    },
...
}